fixed table creation

+++ src/db_sql.c
@@ -81,18 +81,33 @@ int db_writespec_sql(db_config* dbconf){
 					    390 + length of table
 					    name should be enough.
 					 */
-  /* We'll drop table and then recreate it */
+  /* We'll drop table and then recreate it if it exists */
 
-  /* drop the table */
-  sprintf(s, "DROP TABLE %s;", ((psql_data*)dbconf->db_out)->table);
-  res = PQexec(((psql_data*)dbconf->db_out)->conn, s);
-  if ( _db_check_result(((psql_data*)dbconf->db_out)->conn, res, s) == 0 )
+  /* check if the table exists already */
+  sprintf(s, "SELECT * FROM pg_class WHERE relname = '%s'", 
+             ((psql_data*)conf->db_out)->table);
+  res = PQexec(((psql_data*)conf->db_out)->conn, s);
+  if ( _db_check_result(((psql_data*)conf->db_out)->conn, res, s) == 0 ) {
     ret = RETFAIL;
-
+  }
+  table_exists = PQntuples(res) == 1 ? 1 : 0;
   PQclear(res);
 
-  /* Now we need to create the table */
   *s = '\0';  /* reset query string */
+
+  if (table_exists == 1) {
+    /* drop the table */
+    sprintf(s, "DROP TABLE %s;", ((psql_data*)dbconf->db_out)->table);
+    res = PQexec(((psql_data*)dbconf->db_out)->conn, s);
+    if ( _db_check_result(((psql_data*)dbconf->db_out)->conn, res, s) == 0 )
+      ret = RETFAIL;
+
+    PQclear(res);
+
+    *s = '\0';  /* reset query string */
+  }
+
+  /* Now we need to create the table */
   s = strcat(s, "CREATE TABLE ");
   s = strcat(s, ((psql_data*)dbconf->db_out)->table);
   s = strcat(s, "(");
